% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{expandramp}
\alias{expandramp}
\title{Create color ramp palette from fixed color objects}
\usage{
expandramp(x, n)
}
\arguments{
\item{x}{A \code{data.frame} object with two columns: \code{color} for hexadecimal color values, \code{z} for their position.}

\item{n}{Single integer number}
}
\value{
A \code{list}-class object, with three elements: \code{$col} hexadecimal color values, \code{$mid} z-values of midpoints (one for every color, \code{$breaks} separator borders between color values.
}
\description{
Function to create ramp palettes from fixed color positions
}
