% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{ramps}
\alias{ramps}
\alias{gradinv}
\title{Colour gradient ramps}
\usage{
gradinv(n)
}
\arguments{
\item{n}{(\code{numeric}) Number of different colors to generate from the palette}
}
\value{
A function producing a colour gradient ramp.
}
\description{
The object contains functions produced by the \code{\link[grDevices:colorRamp]{colorRampPalette}} function.
}
\details{
\code{showPal} can be used to display the available palettes. You can use \code{pal = "all"} or \code{pal=""} if you want to look at all the available palettes. 
You can also view single palettes individually. The following colour palettes are implemented:
\itemize{
\item \code{gradinv()}: inverse heatmap.
}
}
