% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areas.R
\name{areas}
\alias{areas}
\alias{areas,sfc-method}
\title{Calculate areas inside an arbitrarily defined unit}
\usage{
areas(x, ...)

\S4method{areas}{sfc}(x, lat = NULL, plot = FALSE, s2 = FALSE, ...)
}
\arguments{
\item{x}{A spatial object, either an sf, or sfc, or SpatRaster or RasterArray class object.}

\item{...}{additional arguments passed to class-specific methods.}

\item{lat}{Latitudinal extent. A numeric vector of two values.}

\item{s2}{Should S2 be turned off for the area calculations?}
}
\value{
An numeric area.
}
\description{
Calculate areas inside an arbitrarily defined unit
}
